Moving some definitions around, mostly in splitLinesCollection.ts.
